requirements.txt
langchain-openai == 0.1.8
python-dotenv == 1.0.1
langchain == 0.2.1
langchain-community == 0.2.1
langchain-anthropic == 0.1.15
langchain-google-genai == 1.0.5
sentence-transformers == 3.0.0
bs4 == 0.0.2
langchainhub == 0.1.18